requirements.txt
pyrogram
tgcrypto
google-api-python-client
google-auth
requests
python-dotenv
tinydb
aiofiles
flask
